# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lingo API",
    "summary": "Backend for a Duolingo-inspired language-learning app.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in with email (or username) and password",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the email/password is wrong (NOT_AUTHENTICATED, INVALID_CREDENTIALS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request (VALIDATION_ERROR, INVALID_ANSWER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create an account and sign in",
        "operationId": "signup_api_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state (e.g. OUT_OF_HEARTS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request (VALIDATION_ERROR, INVALID_ANSWER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign out (the client then discards its token)",
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in, or the email/password is wrong (NOT_AUTHENTICATED, INVALID_CREDENTIALS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request (VALIDATION_ERROR, INVALID_ANSWER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lessons/{lesson_id}": {
      "get": {
        "tags": [
          "lessons"
        ],
        "summary": "Lesson exercises for play (never includes solutions)",
        "operationId": "get_lesson_api_lessons__lesson_id__get",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (e.g. LESSON_LOCKED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request (VALIDATION_ERROR, INVALID_ANSWER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{lesson_id}/attempts": {
      "post": {
        "tags": [
          "lessons"
        ],
        "summary": "Start a lesson attempt, or resume the active one",
        "description": "201 when a new attempt is created, 200 when the in-progress attempt is resumed. Optional body `{\"mode\": \"legendary\"}` starts a Legendary challenge on a completed lesson.",
        "operationId": "start_attempt_api_lessons__lesson_id__attempts_post",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StartAttemptIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "200": {
            "description": "Resumed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (e.g. LESSON_LOCKED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state (e.g. OUT_OF_HEARTS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request (VALIDATION_ERROR, INVALID_ANSWER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{lesson_id}/check": {
      "post": {
        "tags": [
          "lessons"
        ],
        "summary": "Check one answer (idempotent per submission_id)",
        "operationId": "check_answer_api_lessons__lesson_id__check_post",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckAnswerIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckAnswerOut"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state (e.g. OUT_OF_HEARTS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request (VALIDATION_ERROR, INVALID_ANSWER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttemptMode": {
        "type": "string",
        "enum": [
          "standard",
          "legendary"
        ],
        "title": "AttemptMode"
      },
      "AttemptOut": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "mode": {
            "$ref": "#/components/schemas/AttemptMode"
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "Deadline for timed challenges; null otherwise."
          },
          "mistake_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mistake Limit",
            "description": "Mistakes allowed in a challenge; null otherwise."
          },
          "solved_exercise_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Solved Exercise Ids",
            "description": "Exercises already answered correctly."
          },
          "mistakes": {
            "type": "integer",
            "title": "Mistakes"
          },
          "total_exercises": {
            "type": "integer",
            "title": "Total Exercises"
          },
          "hearts": {
            "$ref": "#/components/schemas/HeartsOut"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "lesson_id",
          "mode",
          "status",
          "started_at",
          "expires_at",
          "mistake_limit",
          "solved_exercise_ids",
          "mistakes",
          "total_exercises",
          "hearts"
        ],
        "title": "AttemptOut"
      },
...
      "AttemptStatus": {
        "type": "string",
        "enum": [
          "in_progress",
          "completed",
          "failed"
        ],
        "title": "AttemptStatus"
      },
      "CheckAnswerIn": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Attempt Id"
          },
          "exercise_id": {
            "type": "integer",
            "title": "Exercise Id"
          },
          "submission_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Submission Id",
            "description": "Client-generated id (e.g. UUID). Retrying with the same id is safe."
          },
          "answer": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MultipleChoiceAnswer"
              },
              {
                "$ref": "#/components/schemas/WordBankAnswer"
              },
              {
                "$ref": "#/components/schemas/MatchPairsAnswer"
              },
              {
                "$ref": "#/components/schemas/FillBlankAnswer"
              },
              {
                "$ref": "#/components/schemas/TypeAnswerAnswer"
              }
            ],
            "title": "Answer",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "fill_blank": "#/components/schemas/FillBlankAnswer",
                "match_pairs": "#/components/schemas/MatchPairsAnswer",
                "multiple_choice": "#/components/schemas/MultipleChoiceAnswer",
                "type_answer": "#/components/schemas/TypeAnswerAnswer",
                "word_bank": "#/components/schemas/WordBankAnswer"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "exercise_id",
          "submission_id",
          "answer"
        ],
        "title": "CheckAnswerIn"
      },
      "CheckAnswerOut": {
        "properties": {
          "submission_id": {
            "type": "string",
            "title": "Submission Id"
          },
          "exercise_id": {
            "type": "integer",
            "title": "Exercise Id"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "heart_lost": {
            "type": "boolean",
            "title": "Heart Lost",
            "description": "Whether this answer cost a heart."
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer",
            "description": "Revealed only after the answer has been checked."
          },
          "reveal": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MultipleChoiceReveal"
              },
              {
                "$ref": "#/components/schemas/WordBankReveal"
              },
              {
                "$ref": "#/components/schemas/MatchPairsReveal"
              },
              {
                "$ref": "#/components/schemas/FillBlankReveal"
              },
              {
                "$ref": "#/components/schemas/TypeAnswerReveal"
              }
            ],
            "title": "Reveal",
            "description": "Structured correct answer, for highlighting in the UI.",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "fill_blank": "#/components/schemas/FillBlankReveal",
                "match_pairs": "#/components/schemas/MatchPairsReveal",
                "multiple_choice": "#/components/schemas/MultipleChoiceReveal",
                "type_answer": "#/components/schemas/TypeAnswerReveal",
                "word_bank": "#/components/schemas/WordBankReveal"
              }
            }
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Soft feedback, e.g. an accent reminder."
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "hearts": {
            "$ref": "#/components/schemas/HeartsOut"
          },
          "attempt": {
            "$ref": "#/components/schemas/AttemptProgressOut"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "exercise_id",
          "is_correct",
          "heart_lost",
          "correct_answer",
          "reveal",
          "note",
          "explanation",
          "hearts",
          "attempt"
        ],
        "title": "CheckAnswerOut"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "LESSON_LOCKED"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "This lesson is locked."
            ]
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "FillBlankAnswer": {
        "properties": {
          "type": {
            "type": "string",
            "const": "fill_blank",
            "title": "Type",
            "default": "fill_blank"
          },
          "text": {
            "type": "string",
            "maxLength": 100,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "FillBlankAnswer"
      },
...
      "FillBlankExerciseOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "type": {
            "type": "string",
            "const": "fill_blank",
            "title": "Type"
          },
          "content": {
            "$ref": "#/components/schemas/FillBlankContent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "position",
          "prompt",
          "type",
          "content"
        ],
        "title": "FillBlankExerciseOut"
      },
...
      "HeartsOut": {
        "properties": {
          "current": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Current"
          },
          "max": {
            "type": "integer",
            "title": "Max"
          },
          "next_heart_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Heart At",
            "description": "When the next heart regenerates; null if full."
          },
          "regen_minutes": {
            "type": "integer",
            "title": "Regen Minutes"
          },
          "refill_cost_gems": {
            "type": "integer",
            "title": "Refill Cost Gems"
          }
        },
        "type": "object",
        "required": [
          "current",
          "max",
          "next_heart_at",
          "regen_minutes",
          "refill_cost_gems"
        ],
        "title": "HeartsOut"
      },
...
      "LessonOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "skill_id": {
            "type": "integer",
            "title": "Skill Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "xp_reward": {
            "type": "integer",
            "title": "Xp Reward"
          },
          "exercises": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/MultipleChoiceExerciseOut"
                },
                {
                  "$ref": "#/components/schemas/WordBankExerciseOut"
                },
                {
                  "$ref": "#/components/schemas/MatchPairsExerciseOut"
                },
                {
                  "$ref": "#/components/schemas/FillBlankExerciseOut"
                },
                {
                  "$ref": "#/components/schemas/TypeAnswerExerciseOut"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "fill_blank": "#/components/schemas/FillBlankExerciseOut",
                  "match_pairs": "#/components/schemas/MatchPairsExerciseOut",
                  "multiple_choice": "#/components/schemas/MultipleChoiceExerciseOut",
                  "type_answer": "#/components/schemas/TypeAnswerExerciseOut",
                  "word_bank": "#/components/schemas/WordBankExerciseOut"
                }
              }
            },
            "type": "array",
            "title": "Exercises"
          }
        },
        "type": "object",
        "required": [
          "id",
          "skill_id",
          "title",
          "xp_reward",
          "exercises"
        ],
        "title": "LessonOut",
        "description": "Everything needed to render a lesson — and nothing that reveals an answer."
      },
...
      "LoginIn": {
        "properties": {
          "identifier": {
            "type": "string",
            "maxLength": 254,
            "minLength": 1,
            "title": "Identifier",
            "description": "Email address or username."
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "password"
        ],
        "title": "LoginIn"
      },
      "MatchPairsAnswer": {
        "properties": {
          "type": {
            "type": "string",
            "const": "match_pairs",
            "title": "Type",
            "default": "match_pairs"
          },
          "pairs": {
            "items": {
              "$ref": "#/components/schemas/SubmittedPair"
            },
            "type": "array",
            "maxItems": 6,
            "minItems": 1,
            "title": "Pairs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pairs"
        ],
        "title": "MatchPairsAnswer"
      },
...
      "MatchPairsExerciseOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "type": {
            "type": "string",
            "const": "match_pairs",
            "title": "Type"
          },
          "content": {
            "$ref": "#/components/schemas/MatchPairsContent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "position",
          "prompt",
          "type",
          "content"
        ],
        "title": "MatchPairsExerciseOut"
      },
...
      "MultipleChoiceAnswer": {
        "properties": {
          "type": {
            "type": "string",
            "const": "multiple_choice",
            "title": "Type",
            "default": "multiple_choice"
          },
          "option_id": {
            "type": "string",
            "maxLength": 16,
            "minLength": 1,
            "title": "Option Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "option_id"
        ],
        "title": "MultipleChoiceAnswer"
      },
...
      "MultipleChoiceExerciseOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "type": {
            "type": "string",
            "const": "multiple_choice",
            "title": "Type"
          },
          "content": {
            "$ref": "#/components/schemas/MultipleChoiceContent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "position",
          "prompt",
          "type",
          "content"
        ],
        "title": "MultipleChoiceExerciseOut"
      },
      "MultipleChoiceReveal": {
        "properties": {
          "type": {
            "type": "string",
            "const": "multiple_choice",
            "title": "Type",
            "default": "multiple_choice"
          },
          "correct_option_id": {
            "type": "string",
            "title": "Correct Option Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "correct_option_id"
        ],
        "title": "MultipleChoiceReveal"
      },
...
      "SessionOut": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Send as `Authorization: Bearer <token>` on every request."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "title": "SessionOut"
      },
...
      "SignupIn": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "title": "Email",
            "description": "Becomes the sign-in name; stored lower-case.",
            "examples": [
              "sam@example.com"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "SignupIn"
      },
...
      "StartAttemptIn": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/AttemptMode",
            "default": "standard"
          }
        },
        "type": "object",
        "title": "StartAttemptIn"
      },
...
      "TypeAnswerAnswer": {
        "properties": {
          "type": {
            "type": "string",
            "const": "type_answer",
            "title": "Type",
            "default": "type_answer"
          },
          "text": {
            "type": "string",
            "maxLength": 300,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TypeAnswerAnswer"
      },
...
      "TypeAnswerExerciseOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "type": {
            "type": "string",
            "const": "type_answer",
            "title": "Type"
          },
          "content": {
            "$ref": "#/components/schemas/TypeAnswerContent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "position",
          "prompt",
          "type",
          "content"
        ],
        "title": "TypeAnswerExerciseOut"
      },
...
      "WordBankAnswer": {
        "properties": {
          "type": {
            "type": "string",
            "const": "word_bank",
            "title": "Type",
            "default": "word_bank"
          },
          "tile_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "minItems": 1,
            "title": "Tile Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tile_ids"
        ],
        "title": "WordBankAnswer"
      },
...
      "WordBankExerciseOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "type": {
            "type": "string",
            "const": "word_bank",
            "title": "Type"
          },
          "content": {
            "$ref": "#/components/schemas/WordBankContent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "position",
          "prompt",
          "type",
          "content"
        ],
        "title": "WordBankExerciseOut"
      },
...
    }
  }
}